+++ examples/loaders/CMakeLists.txt
@@ -10,9 +10,9 @@ if (assimp_FOUND)
     target_link_libraries(assimp_loader PUBLIC threepp assimp::assimp)
 endif()
 
-if (THREEPP_WITH_SVG_FEATURE)
+if (pugixml_FOUND)
     add_executable(svg_loader svg_loader.cpp)
-    target_link_libraries(svg_loader PUBLIC threepp)
+    target_link_libraries(svg_loader PUBLIC threepp pugixml::pugixml)
     if (imgui_FOUND)
         target_link_libraries(svg_loader PRIVATE imgui::imgui)
     endif ()
