+++ include/threepp/loaders/svg/SVGTypes.hpp
